karchive: Explicitly convert enum to int for QString::arg
The current code doesn't compile after
https://github.com/qt/qtbase/commit/0a1c69bba5b5dd26b937ac0a67ab0bdd1c32b221
(released in Qt 6.10.1)

Closes #4294

(cherry picked from commit 64175fc9566a9da75e3d4818b9b3d9f600893f68)

+++ src/karchive/src/kar.cpp
@@ -90,7 +90,7 @@ bool KAr::openArchive(QIODevice::OpenMode mode)
         return true;
     }
     if (mode != QIODevice::ReadOnly && mode != QIODevice::ReadWrite) {
-        setErrorString(tr("Unsupported mode %1").arg(mode));
+        setErrorString(tr("Unsupported mode %1").arg(static_cast<int>(mode)));
         return false;
     }
